requirements.txt
pytorch
torchvision
numpy
matplotlib
h5py
tensorboard
tensorboardX
notebook
ipython
coverage
tdqm
